TVB-2687 Rollback recently introduced dependency between ProjectService -> OperationService TVB-2663 Cleanup undesired dependency from tvb.core.services.operation_service towards tvb.adapters.simulator
