+++ framework_tvb/tvb/adapters/simulator/simulator_adapter.py
@@ -41,80 +41,26 @@ Few supplementary steps are done here:
 
 """
 import json
-
-from tvb.adapters.simulator.simulator_fragments import SimulatorRMFragment, SimulatorSurfaceFragment, \
-    SimulatorStimulusFragment, SimulatorModelFragment, SimulatorIntegratorFragment, SimulatorMonitorFragment, \
-    SimulatorFinalFragment
-from tvb.core.neotraits.view_model import ViewModel, DataTypeGidAttr
-from tvb.datatypes.connectivity import Connectivity
-from tvb.datatypes.cortex import Cortex
-from tvb.datatypes.local_connectivity import LocalConnectivity
-from tvb.datatypes.patterns import SpatioTemporalPattern, StimuliSurface
-from tvb.datatypes.region_mapping import RegionMapping
-from tvb.datatypes.surfaces import CorticalSurface
-from tvb.simulator.coupling import Coupling
-from tvb.simulator.simulator import Simulator
+from tvb.adapters.simulator.simulator_fragments import *
 from tvb.adapters.simulator.coupling_forms import get_ui_name_to_coupling_dict
-from tvb.adapters.datatypes.h5.simulation_history_h5 import SimulationHistory
 from tvb.adapters.datatypes.db.simulation_history import SimulationHistoryIndex
 from tvb.adapters.datatypes.db.region_mapping import RegionMappingIndex, RegionVolumeMappingIndex
 from tvb.adapters.datatypes.db.connectivity import ConnectivityIndex
-from tvb.adapters.datatypes.db.time_series import TimeSeriesIndex, Attr
+from tvb.adapters.datatypes.db.time_series import TimeSeriesIndex
+from tvb.basic.neotraits.api import Attr
+from tvb.core.entities.file.simulator.simulation_history_h5 import SimulationHistory
+from tvb.core.entities.file.simulator.view_model import SimulatorAdapterModel
 from tvb.core.entities.storage import dao
 from tvb.core.adapters.abcadapter import ABCAsynchronous, ABCAdapterForm
 from tvb.core.adapters.exceptions import LaunchException
 from tvb.core.neotraits.forms import DataTypeSelectField, FloatField, SelectField
 from tvb.core.neocom import h5
-
-
-class CortexViewModel(ViewModel, Cortex):
-
-    surface_gid = DataTypeGidAttr(
-        linked_datatype=CorticalSurface
-    )
-
-    local_connectivity = DataTypeGidAttr(
-        linked_datatype=LocalConnectivity,
-        required=Cortex.local_connectivity.required,
-        label=Cortex.local_connectivity.label,
-        doc=Cortex.local_connectivity.doc
-    )
-
-    region_mapping_data = DataTypeGidAttr(
-        linked_datatype=RegionMapping,
-        label=Cortex.region_mapping_data.label,
-        doc=Cortex.region_mapping_data.doc
-    )
-
-
-class SimulatorAdapterModel(ViewModel, Simulator):
-    connectivity = DataTypeGidAttr(
-        linked_datatype=Connectivity,
-        required=Simulator.connectivity.required,
-        label=Simulator.connectivity.label,
-        doc=Simulator.connectivity.doc
-    )
-
-    surface = Attr(
-        field_type=CortexViewModel,
-        label=Simulator.surface.label,
-        default=Simulator.surface.default,
-        required=Simulator.surface.required,
-        doc=Simulator.surface.doc
-    )
-
-    stimulus = DataTypeGidAttr(
-        linked_datatype=SpatioTemporalPattern,
-        label=Simulator.stimulus.label,
-        default=Simulator.stimulus.default,
-        required=Simulator.stimulus.required,
-        doc=Simulator.stimulus.doc
-    )
-
-    history_gid = DataTypeGidAttr(
-        linked_datatype=SimulationHistory,
-        required=False
-    )
+from tvb.core.services.simulator_serializer import SimulatorSerializer
+from tvb.datatypes.cortex import Cortex
+from tvb.datatypes.patterns import StimuliSurface
+from tvb.datatypes.surfaces import CorticalSurface
+from tvb.simulator.coupling import Coupling
+from tvb.simulator.simulator import Simulator
 
 
 class SimulatorAdapterForm(ABCAdapterForm):
@@ -185,9 +131,15 @@ class SimulatorAdapter(ABCAsynchronous):
     def get_form_class(self):
         return SimulatorAdapterForm
 
-    def get_simulator_fragments(self):
+    @staticmethod
+    def get_simulator_fragments():
         return [SimulatorSurfaceFragment, SimulatorRMFragment, SimulatorStimulusFragment, SimulatorModelFragment,
                 SimulatorIntegratorFragment, SimulatorMonitorFragment, SimulatorFinalFragment]
+
+    def load_view_model(self, operation):
+        storage_path = self.file_handler.get_project_folder(operation.project, str(operation.id))
+        input_gid = json.loads(operation.parameters)['gid']
+        return SimulatorSerializer().deserialize_simulator(input_gid, storage_path)
 
     def get_output(self):
         """
