+++ framework_tvb/tvb/tests/framework/core/entities/file/datatypes/neotraits_h5_test.py
@@ -34,13 +34,13 @@ import tvb
 from tvb.basic.neotraits.ex import TraitAttributeError
 from tvb.adapters.datatypes.h5.local_connectivity_h5 import LocalConnectivityH5
 from tvb.adapters.datatypes.h5.projections_h5 import ProjectionMatrixH5
-from tvb.adapters.datatypes.h5.simulation_history_h5 import SimulationHistoryH5, SimulationHistory
 from tvb.adapters.datatypes.h5.structural_h5 import StructuralMRIH5
 from tvb.adapters.datatypes.h5.volumes_h5 import VolumeH5
 from tvb.adapters.datatypes.h5.connectivity_h5 import ConnectivityH5
 from tvb.adapters.datatypes.h5.region_mapping_h5 import RegionMappingH5
 from tvb.adapters.datatypes.h5.sensors_h5 import SensorsH5
 from tvb.adapters.datatypes.h5.surface_h5 import SurfaceH5
+from tvb.core.entities.file.simulator.simulation_history_h5 import SimulationHistoryH5, SimulationHistory
 from tvb.datatypes.local_connectivity import LocalConnectivity
 from tvb.datatypes.projections import ProjectionMatrix
 from tvb.datatypes.connectivity import Connectivity
